📝 updated swagger file
