+++ 2NotyBackend/routes/catalogs.js
@@ -47,17 +47,10 @@ const {
   getMarks,
   updateMark,
   postMark,
-  getSubscriptions,
-  postSubscription,
-  updateSubscription,
-  getSubscribers,
-  postSubscriptor,
-  updateSubscriptor,
   getTypesPay,
   postTypePay,
   updateTypePay,
   getPaymentsMeans,
-  getSubscriptionDetail,
   postPaymentsMeans,
   getPaymentsUser
 } = require("../controllers/catalogs");
@@ -70,12 +63,10 @@ router.get("/getCompanies", getCompanies);
 router.get("/getBills", getBills);
 router.get("/getCategories", getCategories);
 router.get("/getMarks", getMarks);
-router.get("/getSubscriptions", getSubscriptions);
-router.get("/getSubscribers", getSubscribers);
 router.get("/getTypespay", getTypesPay);
 router.get("/getPaymentsmeans", getPaymentsMeans);
 router.get("/getPaymentsuser", getPaymentsUser);
-router.get("/getSubscriptiondetail", getSubscriptionDetail);
+
 
 router.post(
   "/state",
@@ -116,8 +107,7 @@ router.post(
 )
 router.post("/category", postCategory)
 router.post("/mark", postMark)
-router.post("/subscription", postSubscription)
-router.post("/subscriptor", postSubscriptor)
+
 router.post("/typepay", postTypePay)
 router.post("/paymentsmeans", postPaymentsMeans)
 router.post("/saveImage", fileUpload, (req, res) => {
@@ -136,8 +126,8 @@ router.put("/updateUser", updateUser);
 router.put("/updateCompany", updateCompany);
 router.put("/updateCategory", updateCategory);
 router.put("/updateMark", updateMark);
-router.put("/updateSubscription", updateSubscription);
-router.put("/updateSubscriptor", updateSubscriptor);
+
+
 router.put("/updateTypepay", updateTypePay);
 /* router.put("/activateCountry/:id", activateCountry);
 router.put("/activateState/:id", activateState); */
