+++ 2NotyBackend/routes/subcribers.js
@@ -1,0 +1,15 @@
+/*
+Rutas para Suscripciones
+host + /api/subcribers
+ */
+const router = require("express").Router();
+const {
+    getSubscribers,
+    updateSubscribers,
+    postSubscribers
+} = require("../controllers/subscribers");
+
+router.get("/", getSubscribers);
+router.put("/updateSubscribers/", updateSubscribers);
+router.post("/insertSubscribers/", postSubscribers)
+module.exports = router;
